backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
geoalchemy2==0.14.3
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.18.0
google-auth==2.27.0
numpy==1.26.3
pandas==2.2.3
shapely==2.0.2
geojson==3.1.0
python-dotenv==1.0.0
jinja2==3.1.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# AI/ML dependencies
tensorflow>=2.13.0
scikit-learn>=1.3.0
h5py>=3.9.0

# Earth Engine
earthengine-api>=0.1.400
geemap>=0.32.0
